move product and price models to server module
