+++ server/src/main/java/GroceryFamily/GroceryMom/repository/PriceRepository.java
@@ -1,6 +1,6 @@
 package GroceryFamily.GroceryMom.repository;
 
-import GroceryFamily.GroceryElders.model.Price;
+import GroceryFamily.GroceryMom.model.Price;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PriceRepository extends JpaRepository<Price, String> {}
