+++ server/src/main/java/GroceryFamily/GroceryMom/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package GroceryFamily.GroceryMom.repository;
 
-import GroceryFamily.GroceryElders.model.Product;
+import GroceryFamily.GroceryMom.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, String> {}
