+++ server/src/test/java/GroceryFamily/GroceryMom/service/ProductServiceTest.java
@@ -18,6 +18,6 @@ class ProductServiceTest {
 
     @Test
     void test() {
-
+        // todo: implement
     }
 }
